requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
django-cors-headers==3.5.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-react-templatetags==6.0.2  # via -r requirements.in
django-restql==0.11.0     # via -r requirements.in
django==3.0               # via -r requirements.in, django-cors-headers, django-filter, django-react-templatetags, django-restql, djangorestframework
djangorestframework==3.12.2  # via -r requirements.in, django-restql
factory-boy==3.1.0        # via -r requirements.in
faker==4.17.1             # via factory-boy
gunicorn==20.0.4          # via -r requirements.in
pypeg2==2.15.2            # via django-restql
python-dateutil==2.8.1    # via faker
pytz==2020.4              # via django
pyyaml==5.3.1             # via -r requirements.in
six==1.15.0               # via python-dateutil
sqlparse==0.4.1           # via django
text-unidecode==1.3       # via faker
whitenoise==5.2.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
